updated
git-svn-id: svn+ssh://neptune.elsevier.com/usr/local/subversion/repos/baja/trunk@5775 aa8a525f-4958-0410-8862-b993faadc08d

+++ eijava/eilib/src/org/ei/data/bd/loadtime/BaseTableRecord.java
@@ -42,7 +42,7 @@ public class BaseTableRecord
    	public static final BaseTableRecord AFFILIATION = new BaseTableRecord("AFFILIATION", new Integer(4000), "cpx");
    	public static final BaseTableRecord AFFILIATION_1 = new BaseTableRecord("AFFILIATION_1", new Integer(4000), "cpx");
    	public static final BaseTableRecord CODEN = new BaseTableRecord("CODEN", new Integer(6), "cpx");
-   	public static final BaseTableRecord ISSUE = new BaseTableRecord("ISSUE", new Integer(65), "cpx");
+   	public static final BaseTableRecord ISSUE = new BaseTableRecord("ISSUE", new Integer(128), "cpx");
    	public static final BaseTableRecord TITLELANGUAGE= new BaseTableRecord("TITLELANGUAGE", new Integer(32), "cpx");
    	public static final BaseTableRecord CLASSIFICATIONCODE = new BaseTableRecord("CLASSIFICATIONCODE", new Integer(768), "cpx");
    	public static final BaseTableRecord CONTROLLEDTERM = new BaseTableRecord("CONTROLLEDTERM", new Integer(4000), "cpx");
@@ -61,7 +61,7 @@ public class BaseTableRecord
    	public static final BaseTableRecord ISSN= new BaseTableRecord("ISSN", new Integer(9), "cpx");
    	public static final BaseTableRecord EISSN = new BaseTableRecord("EISSN", new Integer(13), "cpx");
    	public static final BaseTableRecord ISBN= new BaseTableRecord("ISBN", new Integer(1000), "cpx");
-   	public static final BaseTableRecord VOLUME= new BaseTableRecord("VOLUME", new Integer(32), "cpx");
+   	public static final BaseTableRecord VOLUME= new BaseTableRecord("VOLUME", new Integer(128), "cpx");
    	public static final BaseTableRecord PAGE= new BaseTableRecord("PAGE", new Integer(120), "cpx");
    	public static final BaseTableRecord PAGECOUNT= new BaseTableRecord("PAGECOUNT", new Integer(32), "cpx");
    	public static final BaseTableRecord ARTICLENUMBER  = new BaseTableRecord("ARTICLENUMBER", new Integer(64), "cpx");
